develop to main (#22)
* CRO-59 sidemission

* CRO-59 sidemission

* CRO-59 sidemission

* CRO-59 sidemission

* CRO-59 Added C# script for generating heatmap

* CRO-59 Added routes for getting and posting detected detected devices

* CRO-59 Less bad code maybe

* CRO-41 WIP. Added new Service

* CRO-41 Calculate points

* CRO-41 post for rasps

* Update CrowdedBackend/CrowdedBackend/Services/CalculatePositions/CircleUtils.cs


* Update CrowdedBackend/CrowdedBackend/Controllers/DetectedDevicesController.cs


* Cro 41/calculate location2 (#21)

* CRO-41 fix

* CRO-41 updated database

---------

+++ CrowdedBackend/CrowdedBackend/Models/RaspberryPi.cs
@@ -12,8 +12,8 @@ public class RaspberryPi
     [ForeignKey("Venue")]
     public int VenueID { get; set; }
 
-    public int raspX { get; set; }
-    public int raspY { get; set; }
+    public double raspX { get; set; }
+    public double raspY { get; set; }
 
     // Navigation property
     public Venue Venue { get; set; }
